+++ src/middleware.js
@@ -3,19 +3,34 @@ import { jwtVerify } from 'jose';
 
 const JWT_SECRET = process.env.JWT_SECRET;
 
-// convert the secret key to a Uint8Array for jose
+// Convert the secret key to a Uint8Array for jose
 const secret = new TextEncoder().encode(JWT_SECRET);
 
 export async function middleware(req) {
     const { cookies, nextUrl } = req;
     const token = cookies.get('authToken')?.value;
 
-    // allow access to /login page and /api/user/login route without authentication
-    if (nextUrl.pathname === '/login' || nextUrl.pathname === '/api/user/login') {
+    // Check if the user is accessing the /login page
+    if (nextUrl.pathname === '/login') {
+        // If token exists and is valid, redirect to the homepage
+        if (token) {
+            try {
+                await jwtVerify(token, secret);
+                return NextResponse.redirect(new URL('/', nextUrl.origin));
+            } catch (error) {
+                console.error('Invalid or expired token:', error);
+            }
+        }
+        // Allow access to /login if no valid token exists
         return NextResponse.next();
     }
 
-    // check on any other url/route
+    // Allow access to /api/user/login route without authentication
+    if (nextUrl.pathname === '/api/user/login') {
+        return NextResponse.next();
+    }
+
+    // For other routes, check for authentication
     if (!token) {
         return NextResponse.redirect(new URL('/login', nextUrl.origin));
     }
@@ -29,10 +44,10 @@ export async function middleware(req) {
     }
 }
 
-// middleware to pages and API routes (excluding /api/user/login)
+// Middleware configuration for pages and API routes (excluding specific paths)
 export const config = {
     matcher: [
-        '/((?!_next/static|_next/image|favicon.ico|login|api/user/login).*)',
+        '/((?!_next/static|_next/image|favicon.ico|api/user/login).*)',
         '/api/:path*',
     ],
 };
