+++ src/pages/_app.js
@@ -1,3 +1,12 @@
+import Header from "@/components/globalUI/header";
+
 export default function App({ Component, pageProps }) {
-    return <Component {...pageProps} />;
+    return (
+        <>
+            <Header />
+            <Component {...pageProps} />
+        </>
+    )
+
+
 }
