+++ src/pages/api/file/retrieve.js
@@ -1,7 +1,6 @@
 import S3Client from "@/utils/S3Client";
 import jwt from 'jsonwebtoken';
 import { parse } from 'cookie';
-import { validateToken } from "@/utils/validateToken";
 
 const JWT_KEY = process.env.JWT_SECRET;
